# control-plane/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Control-plane is a central microservice of Service Mesh which is responsible for all the gateways configuration. For more information, visit our Documentation(https://github.com/Netcracker/qubership-core-control-plane/tree/main/README.md).",
        "title": "Control Plane API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v3/versions/registry": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Microservice Versions from BG versions registry. Not every combination of the request params allowed,",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Get Microservice Versions from BG versions registry",
                "operationId": "GetMicroserviceVersions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "initialVersion",
                        "name": "initialVersion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "serviceName",
                        "name": "serviceName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.VersionInRegistry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register Microservice Versions in BG versions registry.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Register Microservice Versions in BG versions registry.",
                "operationId": "PostMicroserviceVersions",
                "parameters": [
                    {
                        "description": "ServicesVersionPayload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServicesVersionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v3/versions/registry/services": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Microservice Versions from BG versions registry.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Delete Microservice Versions from BG versions registry.",
                "operationId": "DeleteMicroserviceVersions",
                "parameters": [
                    {
                        "description": "ServicesVersionPayload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServicesVersionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Microservice": {
            "type": "object",
            "properties": {
                "cluster": {
                    "type": "string"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "dto.ServicesVersionPayload": {
            "type": "object",
            "properties": {
                "exists": {
                    "description": "Exists indicates that it is CREATE request payload (not DELETE request). Field is optional and by default\nrequest considered to be CREATE. Only explicit `false` value of this field will cause request to be\ninterpreted as DELETE.",
                    "type": "boolean"
                },
                "namespace": {
                    "type": "string"
                },
                "overridden": {
                    "type": "boolean"
                },
                "services": {
                    "description": "Services contains family names of the microservices being deployed",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "description": "Version holds blue-green version of microservices being registered (or deleted).\nEmpty version field means that deploy is performed in rolling mode =\u003e ACTIVE version should be used.",
                    "type": "string"
                }
            }
        },
...
        "dto.VersionInRegistry": {
            "type": "object",
            "properties": {
                "clusters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Microservice"
                    }
                },
                "stage": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}